+++ SMV/source/shared/string_util.c
@@ -156,10 +156,10 @@ int randint(int min, int max){
   int return_val;
 
   if (min>max){
-    return_val = max+((min-max+1)*rand()/(RAND_MAX+1.0));
+    return_val = max+((min-max+1)*(float)rand()/(float)(RAND_MAX+1));
   }
   else{
-    return_val = min+((max-min+1)*rand()/(RAND_MAX+1.0));
+    return_val = min+((max-min+1)*(float)rand()/(float)(RAND_MAX+1));
   }
   return return_val;
 }
@@ -564,7 +564,7 @@ int match(char *buffer, const char *key){
 int match_upper(char *buffer, const char *key){
   size_t lenbuffer;
   size_t lenkey;
-  int i;
+  size_t i;
 
   lenkey=strlen(key);
   trim(buffer);
